setup new gui elements
